Add files via upload
This was the IOT project to segregate the moisture waste and dry waste separately. This is widely used in Medical fields,hospitals to avoid unneccsary smells like blood or something else. This segregation provides a lot of benifits to the users not only in the medical fields and also all used waste collecting places like house,industry etc...thus avoids a unneccsary consequences in the atmosphere